Fix potential int to float conversion error

+++ Kumi.Game/Charts/Timings/TimeSignature.cs
@@ -21,5 +21,5 @@ public class TimeSignature
         Denominator = denominator;
     }
     
-    public float NoteSubdivision => Numerator / Denominator;
+    public float NoteSubdivision => (float)Numerator / Denominator;
 }
